refactor(automation): merge tiny modules into siblings
Consolidate the former work_report helper into _review_utils and the secret-file constants into pr_manager. Move ReviewerProtocol into protocol.py while preserving _interfaces.py for the broader protocol contracts that now live on main.

Closes #1442

Implemented-By: claude-opus-4-8
