+++ hephaestus/automation/planner.py
@@ -39,6 +39,7 @@ from ._review_utils import (
     close_issue_as_covered,
     find_merged_closing_pr,
     find_pr_for_issue,
+    work_report_context,
 )
 from .advise_runner import advise_skipped, ensure_mnemosyne, run_advise
 from .agent_config import (
@@ -61,7 +62,6 @@ from .prompts import get_advise_prompt_builder
 from .review_state import is_plan_review_go
 from .session_naming import AGENT_ADVISE
 from .status_tracker import StatusTracker
-from .work_report import work_report_context
 
 __all__ = ["MAX_REVIEW_ITERATIONS", "Planner", "main"]
