+++ hephaestus/automation/protocol.py
@@ -20,7 +20,7 @@ here per issue #801 (tracking #708).
 
 from __future__ import annotations
 
-from typing import Final
+from typing import Any, Final, Protocol, runtime_checkable
 
 PLAN_COMMENT_MARKER: Final[str] = "# Implementation Plan"
 """Heading the planner writes at the top of the single plan comment."""
@@ -36,4 +36,19 @@ intentional-design finding (e.g. an abstract method's ``NotImplementedError``)
 cannot stack duplicate threads across runs. Part of the wire protocol — both the
 validator's resolve-reply and the reviewer's dedup match on this exact string."""
 
-__all__ = ["PLAN_COMMENT_MARKER", "PLAN_REVIEW_PREFIX", "WONT_FIX_MARKER"]
+
+@runtime_checkable
+class ReviewerProtocol(Protocol):
+    """Structural contract satisfied by all four reviewer classes.
+
+    Verified: PRReviewer.run (pr_reviewer.py:396),
+              AddressReviewer.run (address_review.py:350),
+              AuditReviewer.run (audit_reviewer.py:197),
+              PlanReviewer.run (plan_reviewer.py:99).
+    """
+
+    def run(self) -> Any:
+        """Execute the reviewer and return its result."""
+
+
+__all__ = ["PLAN_COMMENT_MARKER", "PLAN_REVIEW_PREFIX", "WONT_FIX_MARKER", "ReviewerProtocol"]
